refactor: bridge inputs
