+++ src/views/Disclaimer.vue
@@ -1,0 +1,12 @@
+<template>
+  <!-- TODO: -->
+  <div></div>
+</template>
+
+<script lang="ts">
+import { defineComponent } from 'vue'
+
+export default defineComponent({
+  // components: { },
+})
+</script>
